Refactor out xcom constants from models
