+++ airflow/providers/amazon/aws/transfers/google_api_to_s3.py
@@ -23,9 +23,9 @@ import sys
 from typing import TYPE_CHECKING, Sequence
 
 from airflow.models import BaseOperator, TaskInstance
-from airflow.models.xcom import MAX_XCOM_SIZE, XCOM_RETURN_KEY
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
 from airflow.providers.google.common.hooks.discovery_api import GoogleDiscoveryApiHook
+from airflow.utils.xcom import MAX_XCOM_SIZE, XCOM_RETURN_KEY
 
 if TYPE_CHECKING:
     from airflow.utils.context import Context
