+++ tests/providers/amazon/aws/transfers/test_google_api_to_s3.py
@@ -23,9 +23,9 @@ import pytest
 
 from airflow import models
 from airflow.configuration import conf
-from airflow.models.xcom import MAX_XCOM_SIZE
 from airflow.providers.amazon.aws.transfers.google_api_to_s3 import GoogleApiToS3Operator
 from airflow.utils import db
+from airflow.utils.xcom import MAX_XCOM_SIZE
 
 
 class TestGoogleApiToS3:
